requirements.txt
unsloth
transformers
datasets
accelerate
peft
trl
torch
pandas
scikit-learn
pyyaml
